requirements.txt
mecab-python==0.993
oauthlib==0.6.3
pep8==1.5.7
requests==2.1.0
requests-oauthlib==0.4.0
twython==3.1.2
wsgiref==0.1.2
